Universal RP installed and Prefabs Done
I put the borders that create the shadows on all prefab rooms. The borders are inside a game object in each prefab. I also changed the material in the prefabs for it to be affected by light. Finally, I also changed the material of all other prefabs that need to be affected by light, plus added a blue light to the projectiles