Fully Complete Intro + Visual Novel System WIP
Added a lot of stuff. I added a Scene which is the Intro of the game. It is fully completed. I added the Visual Novel system for story telling and other functions in the game. For Dash, I added Chloe's new sprites to animate in the resources. Will work on finishing some functions and the dialogue in the main VN system then make the endings