New Animations for the player
So, added idle, walking and attacking animations using Chloe's sprites. They are smooth animations and work well overall. Only issue would be the attacking animation might continue for a small fraction of a second after attacking then moving immediately. A solution for this would be to decrease more the attacking animation duration, however, it might have the unintended effect of making it seems Chloe is shooting too fast