Solved Bug of Map Generation and Adjusted Lights
Solved the issue when generating the map with rooms that were not affected by light. I also changed the intensity, radius and fallout of lights to give a more sinister and dark environment. Lastly, I changed the background colour of the camera to black, so you will not be able to see pitfalls in blue anymore and use that information to map mentally the maze. This is only visible when playing the game