Added Tutorial + Tutorial VN + New decorations
I created a tutorial room. The player will spawn there. A visual novel story will play out as the player advance. To exit the room, the player will have to react the end and interact with the Crystal Nexus. It will teleport her to the main maze. To change the Coords to where she teleports, simply change the coords in the Vector3 inside INT_TutorialTeleporter Script. Also, I got new sprites we can use to decorate the map and give it a more sinister ambience. I already made several prehabs of it in a folder called Decorations. The package with all the stuff is inside a folder called RafaelMatos in Assets/RafaelMatos